feat(cli): add CLI infrastructure + stitch, prepare-data, generate-data commands
Create src/speculators/cli/ package with Pipeline/Tools help groups per
RFC #906. Move stitch, prepare-data, and generate-data script logic into
package modules, convert argparse interfaces to typer, and wire as
unified CLI subcommands. Original scripts become backward-compat shims
with deprecation warnings.

Part of rahul-tuli/speculators#28
